rag_engine/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv

pymupdf==1.24.10
python-pptx==1.0.2
Pillow==10.4.0
requests==2.32.3
frontend==0.0.3
llama-index-core==0.10.58
llama-index-readers-file==0.1.30
llama-index-embeddings-huggingface
llama-index-llms-huggingface==0.2.8
llama-index-vector-stores-milvus==0.1.20
llama-index-llms-llama-cpp
pymilvus==2.4.4
torch
transformers